fix: Move delete recs views to the correct folder.
