+++ src/tools/recommendation-views/delete-recommendation-view.test.ts
@@ -1,7 +1,7 @@
 import { pathEncode } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
 import { requestsInOrder, testTool } from "../utils/testing";
-import tool from "./delete-recommendation-view";
+import tool from "../recommendation-views/delete-recommendation-view";
 
 testTool(
   tool,
