+++ src/tools/recommendation-views/delete-recommendation-view.ts
