+++ src/tools/recommendation-views/index.ts
@@ -2,3 +2,4 @@ import "./list-recommendation-views";
 import "./get-recommendation-view";
 import "./create-recommendation-view";
 import "./update-recommendation-view";
+import "./delete-recommendation-view";
