+++ src/tools/recommendations/index.ts
@@ -1,4 +1,3 @@
-import "./delete-recommendation-view";
 import "./get-recommendation-details";
 import "./get-recommendation-resource-details";
 import "./get-recommendation-resources";
